Git 대소문자 구분 안하는 문제 해결
