+++ frontend/src/app/me/favorites/page.tsx
