+++ frontend/src/components/booking/BookingDetails.tsx
