+++ frontend/src/lib/api/admin/AdminHotelApi.ts
