+++ frontend/src/lib/types/FavoriteHotelDto.ts
