+++ frontend/src/lib/types/booking/BookingResponseDetails.ts
