+++ frontend/src/lib/types/booking/BookingResponseSummary.ts
