+++ frontend/src/lib/types/booking/payment/UidResponse.ts
