+++ frontend/src/lib/types/booking/revenue/HotelRevenueResponse.ts
